requirements.txt
anthropic
arxiv
python-dotenv
mcp[cli]
